% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotvarpart.R
\name{plotvarpart}
\alias{plotvarpart}
\title{Create Venn driagrams graph}
\usage{
plotvarpart(
  x,
  x1 = "X1",
  x2 = "X2",
  resi = "Residuals",
  colour = "grey",
  fill = NA,
  grad = "/",
  fill1 = "white",
  fill2 = "black",
  font = "serif"
)
}
\arguments{
\item{x}{One object of class varpart}

\item{x1}{Partition name}

\item{x2}{Other partition name}

\item{resi}{Risduals name}

\item{colour}{Circle colour}

\item{fill}{Circle fill}

\item{grad}{Logical argument to indicate if the graph must be in gradient}

\item{fill1}{Circle fill if it be in gradient. To the minor values}

\item{fill2}{Circle fill if it be in gradient. To the major values}

\item{font}{Change the texts family in graph}
}
\value{
The object of class ggplot
}
\description{
Function to create graphs type Ven diagram using objects of
class varpart. The function has one argument that you to use to make graphs
that show the variance partition in a colors gradient.
}
\examples{
require(vegan)
require(ggplot2)
require(ggforce)
data(mite)
data(mite.env)
data(mite.pcnm)
mod <- varpart(mite, mite.env, mite.pcnm, transfo="hel")
plotvarpart(mod)
}
\author{
Rafael Costa Bastos
}
